requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pillow>=10.2.0
pydantic>=2.5.0
python-multipart>=0.0.6
pytest>=7.4.0
httpx>=0.26.0
